refactor: Split extractor and loader into dedicated packages
