+++ internal/extractor/extractor_test.go
@@ -1,4 +1,4 @@
-package mongo
+package extractor
 
 import (
 	"context"
